LayoutTests/imported/w3c/web-platform-tests/feature-policy/policy-extends-to-sandbox-expected.txt

CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating 'document.featurePolicy.allowedFeatures')


Harness Error (TIMEOUT), message = null

TIMEOUT Feature policy treats opaque origins correctly Test timed out
